add more osd namespacing
