+++ cypress/integration/without_security/check_timeline.js
@@ -39,7 +39,7 @@ describe('check timeline visualization', () => {
     it('.es(*, kibana1=true) should report search error', () => {
       cy.get('[class="view-line"]').type('.es(*, kibana1=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .contains('Timeline request error: undefined Error: Unknown argument to es: kibana1');
@@ -48,7 +48,7 @@ describe('check timeline visualization', () => {
     it('.es(*, kibana=true) should not report search error', () => {
       cy.get('[class="view-line"]').type('.es(*, kibana=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .should('not.exist');
@@ -57,7 +57,7 @@ describe('check timeline visualization', () => {
     it('.es(*, opensearchDashboards=true) should not report search error', () => {
       cy.get('[class="view-line"]').type('.es(*, opensearchDashboards=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .should('not.exist');
@@ -66,7 +66,7 @@ describe('check timeline visualization', () => {
     it('.elasticsearch(*, kibana1=true) should report search error', () => {
       cy.get('[class="view-line"]').type('.elasticsearch(*, kibana1=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .contains('Timeline request error: undefined Error: Unknown argument to es: kibana1');
@@ -75,7 +75,7 @@ describe('check timeline visualization', () => {
     it('.elasticsearch(*, kibana=true) should not report search error', () => {
       cy.get('[class="view-line"]').type('.elasticsearch(*, kibana=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .should('not.exist');
@@ -84,7 +84,7 @@ describe('check timeline visualization', () => {
     it('.elasticsearch(*, opensearchDashboards=true) should not report search error', () => {
       cy.get('[class="view-line"]').type('.elasticsearch(*, opensearchDashboards=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .should('not.exist');
@@ -93,7 +93,7 @@ describe('check timeline visualization', () => {
     it('.opensearch(*, kibana1=true) should report search error', () => {
       cy.get('[class="view-line"]').type('.opensearch(*, kibana1=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .contains('Timeline request error: undefined Error: Unknown argument to es: kibana1');
@@ -102,7 +102,7 @@ describe('check timeline visualization', () => {
     it('.opensearch(*, kibana=true) should not report search error', () => {
       cy.get('[class="view-line"]').type('.opensearch(*, kibana=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .should('not.exist');
@@ -111,7 +111,7 @@ describe('check timeline visualization', () => {
     it('.opensearch(*, opensearchDashboards=true) should not report search error', () => {
       cy.get('[class="view-line"]').type('.opensearch(*, opensearchDashboards=true)');
       cy.get('[data-test-subj="visualizeEditorRenderButton"]').click();
-      cy.waitForLoader();
+      cy.osd.waitForLoader();
       cy.get('[data-test-subj="globalToastList"]')
         .find('[data-test-subj="errorToastMessage"]')
         .should('not.exist');
