+++ cypress/utils/apps/data_explorer/commands.js
@@ -83,7 +83,7 @@ Cypress.Commands.add(
     }
 
     cy.getElementByTestId('saveFilter').click({ force: true });
-    cy.waitForLoader(isEnhancement);
+    cy.osd.waitForLoader(isEnhancement);
   }
 );
 
@@ -118,7 +118,7 @@ Cypress.Commands.add(
 );
 
 Cypress.Commands.add(
-  'updateSaveQuery',
+  'updateSavedQuery',
   (name = '', saveAsNewQuery = false, includeFilters = true, includeTimeFilter = false) => {
     cy.whenTestIdNotFound('saved-query-management-popover', () => {
       cy.getElementByTestId('saved-query-management-popover-button').click();
@@ -155,7 +155,7 @@ Cypress.Commands.add(
   }
 );
 
-Cypress.Commands.add('loadSaveQuery', (name) => {
+Cypress.Commands.add('loadSavedQuery', (name) => {
   cy.getElementByTestId('saved-query-management-popover-button').click();
 
   cy.getElementByTestId('saved-query-management-open-button').click();
@@ -165,7 +165,7 @@ Cypress.Commands.add('loadSaveQuery', (name) => {
   cy.getElementByTestId('open-query-action-button').click({ force: true });
 });
 
-Cypress.Commands.add('clearSaveQuery', () => {
+Cypress.Commands.add('clearSavedQuery', () => {
   cy.whenTestIdNotFound('saved-query-management-popover', () => {
     cy.getElementByTestId('saved-query-management-popover-button').click();
   });
@@ -173,7 +173,7 @@ Cypress.Commands.add('clearSaveQuery', () => {
   cy.getElementByTestId('saved-query-management-clear-button').click();
 });
 
-Cypress.Commands.add('deleteSaveQuery', (name) => {
+Cypress.Commands.add('deleteSavedQuery', (name) => {
   cy.getElementByTestId('saved-query-management-popover-button').click();
 
   cy.getElementByTestId('saved-query-management-open-button').click();
