+++ cypress/utils/dashboards/workspace_plugin/constants.js
